+++ src/public/js/views/consumersSurvey/preView/preView.js
@@ -24,10 +24,11 @@ define([
     'constants/levelConfig',
     'constants/aclRoleIndexes',
     'views/questionnary/editAnswer',
+    'constants/infoMessages'
 ], function (Backbone, _, $, lodash, moment, PreViewTemplate, npsQuestionTemplate, QuestionListTemplate, RespondentsListTemplate, RespondentsFullListTemplate,
              QuestionFullAnswerTemplate, QuestionSingleSelectTemplate, QuestionMultiselectTemplate,
              Model, QuestionModel, QuestionAnswerModel, QuestionCollection, QuestionnaryAnswersCollection, OTHER_CONSTANTS,
-             BaseView, CONTENT_TYPES, d3, LEVEL_CONFIG, ACL_ROLES, EditAnswerView) {
+             BaseView, CONTENT_TYPES, d3, LEVEL_CONFIG, ACL_ROLES, EditAnswerView, INFO_MESSAGES) {
     var preView = BaseView.extend({
         contentType: CONTENT_TYPES.CONSUMER_SURVEY,
 
@@ -111,16 +112,18 @@ define([
         },
 
         deleteAnswer: function (event) {
-            var that = this;
-            var target = $(event.target);
-            var answerId = target.attr('data-id');
-            var model = new QuestionAnswerModel();
+            if (confirm(INFO_MESSAGES.confirmDeleteAnswer[App.currentUser.currentLanguage]))  {
+                var that = this;
+                var target = $(event.target);
+                var answerId = target.attr('data-id');
+                var model = new QuestionAnswerModel();
 
-            model.delete(answerId);
+                model.delete(answerId);
 
-            model.on('answer-deleted', function () {
-                that.$el.find('#respondent-answer-container-' + answerId).remove();
-            });
+                model.on('answer-deleted', function () {
+                    that.$el.find('#respondent-answer-container-' + answerId).remove();
+                });
+            }
         },
 
         duplicateQuestionnary: function () {
